% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{FindCLICFeatures}
\alias{FindCLICFeatures}
\title{Select features for scRNA-seq + scATAC-seq integration using CLIC scores}
\usage{
FindCLICFeatures(
  object,
  species,
  nfeatures = 2000,
  initial_variable_features_num = 5000,
  verbose = TRUE
)
}
\arguments{
\item{object}{A Seurat object or a gene expression counts matrix.}

\item{species}{A character string specifying the species (used to load the corresponding CLIC score file).
Currently, "human" and "mouse" are supported.}

\item{nfeatures}{Number of features to select, equivalent to \code{nfeatures} in Seurat's \code{FindVariableFeatures()}.
Default is 2000.}

\item{initial_variable_features_num}{Number of variable features to compute before applying CLIC selection.
A higher value gives more weight to the CLIC scores. Default is 5000.}

\item{verbose}{Logical. Whether to print progress messages. Default is TRUE.}
}
\value{
A list containing:
\describe{
\item{seurat_obj}{Seurat object containig gene expression data}
\item{use_features}{Character vector of selected features.}
\item{clic_num}{Integer indicating how many CLIC-ranked features were required to reach \code{nfeatures}.
The intersection of the top \code{initial_variable_features_num} variable genes and the top \code{clic_num}
CLIC-ranked genes determines the final feature set.}
}
}
\description{
This function identifies high-confidence features for cross-modal integration
by combining traditional highly variable feature (HVF) selection with
CLIC scores, which quantify the confidence of feature-to-feature links.
}
\details{
In the context of scRNA-seq and scATAC-seq integration, the CLIC score is computed
as the empirical correlation between gene expression (from RNA) and gene activity
(from ATAC) for each gene. Refer to our publication for details on how the CLIC score is computed.

The input can be either a Seurat object (recommended) or a raw counts expression matrix.
If a matrix is provided, it will be converted into a Seurat object internally.
}
\examples{
\dontrun{
out <- FindCLICFeatures(object = obj, species = "human")
use_features <- out$use_features
}

}
\author{
Tomoya Furutani \email{furutomo.49@gmail.com}
}
